Refactor order selector dialog to a separate module.
