+++ picard/ui/options/plugin_execution_order.py
@@ -19,12 +19,10 @@
 # Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
 
 
-from collections import namedtuple
 import sys
 
 from PyQt5 import (
     QtCore,
-    QtGui,
     QtWidgets,
 )
 
@@ -44,13 +42,13 @@ from picard.ui.options import (
     OptionsPage,
     register_options_page,
 )
+from picard.ui.plugin_order_selector import (
+    PluginInformation,
+    display_plugin_order_selector,
+)
 from picard.ui.ui_options_plugin_execution_order import (
     Ui_PluginExecutionOrderOptionsPage,
 )
-from picard.ui.widgets.orderabletableview import OrderableTableView
-
-
-PluginInformation = namedtuple('PluginInformation', ['key', 'name', 'description', 'function_description', 'priority'])
 
 
 class PluginExecutionOrderOptionsPage(OptionsPage):
@@ -97,96 +95,6 @@ class PluginExecutionOrderOptionsPage(OptionsPage):
 
         self.ui.edit_plugin_order.clicked.connect(self.show_execution_order_editor)
 
-        self.setup_editor_dialog()
-
-    def setup_editor_dialog(self):
-        self.order_dialog = QtWidgets.QDialog(self)
-        self.order_dialog.setWindowTitle(_(self.title_text))
-        self.order_dialog.setMinimumWidth(650)
-
-        layout = QtWidgets.QVBoxLayout(self.order_dialog)
-
-        instructions = QtWidgets.QLabel(
-            _(
-                "This displays the order in which the plugin metadata processing functions are executed "
-                "by Picard. You can change the order by moving the functions up or down by selecting "
-                "the funcion to move and then use the up or down button, or by using your mouse to "
-                "drag the function to the desired location in the list."
-            )
-        )
-        instructions.setWordWrap(True)
-        layout.addWidget(instructions)
-
-        self.tableview = OrderableTableView(self.order_dialog)
-
-        self.tableview.model.setHorizontalHeaderLabels(
-            [
-                _('Plugin Name'),
-                _('Processor'),
-                _('Method/Function Called'),
-            ]
-        )
-
-        for idx, width in enumerate([260, 70, 230]):
-            self.tableview.setColumnWidth(idx, width)
-
-        for idx, text in enumerate(
-            [
-                _("The name of the plugin"),
-                _("The type of metadata processor used (album or track)"),
-                _("The method or function within the plugin module that is called"),
-            ]
-        ):
-            self.tableview.model.horizontalHeaderItem(idx).setToolTip(text)
-
-        self.tableview.selectionModel().selectionChanged.connect(self._set_up_down_button_states)
-        layout.addWidget(self.tableview)
-
-        button_layout = QtWidgets.QHBoxLayout()
-
-        # Move Label
-        move_label = QtWidgets.QLabel(_("Move row"))
-        button_layout.addWidget(move_label)
-
-        # Up button
-        self.up_button = QtWidgets.QToolButton()
-        up_icon = self.style().standardIcon(QtWidgets.QStyle.StandardPixmap.SP_TitleBarShadeButton)
-        self.up_button.setIcon(up_icon)
-        self.up_button.setToolTip(_("Move selected plugin up"))
-        self.up_button.clicked.connect(self.tableview.move_row_up)
-        button_layout.addWidget(self.up_button)
-
-        # Down button
-        self.dn_button = QtWidgets.QToolButton()
-        dn_icon = self.style().standardIcon(QtWidgets.QStyle.StandardPixmap.SP_TitleBarUnshadeButton)
-        self.dn_button.setIcon(dn_icon)
-        self.dn_button.setToolTip(_("Move selected plugin down"))
-        self.dn_button.clicked.connect(self.tableview.move_row_down)
-        button_layout.addWidget(self.dn_button)
-
-        # spacer
-        spacer = QtWidgets.QSpacerItem(
-            20, 0, QtWidgets.QSizePolicy.Policy.Expanding, QtWidgets.QSizePolicy.Policy.Minimum
-        )
-        button_layout.addItem(spacer)
-
-        # OK
-        ok_button = QtWidgets.QPushButton(_('OK'))
-        ok_icon = self.style().standardIcon(QtWidgets.QStyle.StandardPixmap.SP_DialogOkButton)
-        ok_button.setIcon(ok_icon)
-        ok_button.clicked.connect(self.order_dialog.accept)
-        button_layout.addWidget(ok_button)
-        ok_button.setDefault(True)  # default selected button
-
-        # Cancel
-        cancel_button = QtWidgets.QPushButton(_('Cancel'))
-        cancel_icon = self.style().standardIcon(QtWidgets.QStyle.StandardPixmap.SP_DialogCancelButton)
-        cancel_button.setIcon(cancel_icon)
-        cancel_button.clicked.connect(self.order_dialog.reject)
-        button_layout.addWidget(cancel_button)
-
-        layout.addLayout(button_layout)
-
     def load(self):
         config = get_config()
         self.plugin_exec_order = config.setting[EXEC_ORDER_KEY]
@@ -219,7 +127,7 @@ class PluginExecutionOrderOptionsPage(OptionsPage):
             msg = QtWidgets.QMessageBox(self)
             msg.setIcon(QtWidgets.QMessageBox.Icon.Warning)
             msg.setText(_("There were no installed metadata processing plugins found."))
-            msg.setWindowTitle(_(self.title_text))
+            msg.setWindowTitle(_("No Data"))
             msg.setWindowModality(QtCore.Qt.WindowModality.ApplicationModal)
             msg.setStandardButtons(QtWidgets.QMessageBox.StandardButton.Ok)
             msg.setDefaultButton(QtWidgets.QMessageBox.StandardButton.Ok)
@@ -227,52 +135,9 @@ class PluginExecutionOrderOptionsPage(OptionsPage):
 
             return
 
-        self.tableview.blockSignals(True)
-        i = self.tableview.model.rowCount()
-        while i > 0:
-            i -= 1
-            self.tableview.model.removeRow(i)
-
-        for plugin in sorted(plugins, key=lambda i: i.priority, reverse=True):
-            plugin: PluginInformation
-            module_name, processor_type, function_name = plugin.key.split(':', 2)
-
-            column1 = QtGui.QStandardItem(plugin.name)
-            column1.setEditable(False)
-            column1.setDropEnabled(False)
-            column1.setToolTip(plugin.description)
-            column1.setData(plugin.key, QtCore.Qt.ItemDataRole.UserRole)
-
-            column2 = QtGui.QStandardItem(processor_type)
-            column2.setEditable(False)
-            column2.setDropEnabled(False)
-            column2.setTextAlignment(QtCore.Qt.AlignCenter)
-
-            column3 = QtGui.QStandardItem(function_name)
-            column3.setEditable(False)
-            column3.setDropEnabled(False)
-            column3.setToolTip(plugin.function_description)
-
-            self.tableview.model.appendRow([column1, column2, column3])
-
-        self.tableview.setCurrentIndex(self.tableview.model.index(0, 0))
-        self.tableview.blockSignals(False)
-
-        self._set_up_down_button_states()
-
-        # Show dialog and process result
-        if self.order_dialog.exec() == QtWidgets.QDialog.DialogCode.Accepted:
-            self.plugin_exec_order = dict()
-            for idx in range(self.tableview.model.rowCount()):
-                item = self.tableview.model.item(idx, 0)
-                key = item.data(QtCore.Qt.ItemDataRole.UserRole)
-                priority = -1 - idx
-                self.plugin_exec_order[key] = priority
-
-    def _set_up_down_button_states(self):
-        row = self.tableview.currentIndex().row()
-        self.up_button.setDisabled(row < 1)
-        self.dn_button.setDisabled(row >= self.tableview.model.rowCount() - 1)
+        new_order, return_state = display_plugin_order_selector(parent=self, plugins=plugins)
+        if return_state:
+            self.plugin_exec_order = new_order
 
 
 register_options_page(PluginExecutionOrderOptionsPage)
